ppc: Remove some SECURE_BOOT defconfigs
These specific configs are missing a number of migrations.  In addition,
they are blocking completion of the now-expired DM_MMC migration as it
requires enabling BLK.

Cc: Priyanka Jain <priyanka.jain@nxp.com>
Cc: Ruchika Gupta <ruchika.gupta@nxp.com>
Cc: Sumit Garg <sumit.garg@nxp.com>
Signed-off-by: Tom Rini <trini@konsulko.com>
---
Note that what I think happened before was that when I removed these
configs I also hit a number of had-been-updated T104* configs.  So
please migrate these ASAP for DM_MMC and note that there's a ton of
v2019.07 migrations due as well.

Also, the T1023* platforms are listed as orphan.